feat: vrgda letter pricing

+++ src/codegen/tables/TilePlayer.sol
@@ -69,8 +69,8 @@ library TilePlayer {
   /** Get value */
   function get(int32 x, int32 y) internal view returns (address value) {
     bytes32[] memory _keyTuple = new bytes32[](2);
-    _keyTuple[0] = bytes32(uint256(uint32((x))));
-    _keyTuple[1] = bytes32(uint256(uint32((y))));
+    _keyTuple[0] = bytes32(uint256(int256(x)));
+    _keyTuple[1] = bytes32(uint256(int256(y)));
 
     bytes memory _blob = StoreSwitch.getField(_tableId, _keyTuple, 0);
     return (address(Bytes.slice20(_blob, 0)));
@@ -79,8 +79,8 @@ library TilePlayer {
   /** Get value (using the specified store) */
   function get(IStore _store, int32 x, int32 y) internal view returns (address value) {
     bytes32[] memory _keyTuple = new bytes32[](2);
-    _keyTuple[0] = bytes32(uint256(uint32((x))));
-    _keyTuple[1] = bytes32(uint256(uint32((y))));
+    _keyTuple[0] = bytes32(uint256(int256(x)));
+    _keyTuple[1] = bytes32(uint256(int256(y)));
 
     bytes memory _blob = _store.getField(_tableId, _keyTuple, 0);
     return (address(Bytes.slice20(_blob, 0)));
@@ -89,8 +89,8 @@ library TilePlayer {
   /** Set value */
   function set(int32 x, int32 y, address value) internal {
     bytes32[] memory _keyTuple = new bytes32[](2);
-    _keyTuple[0] = bytes32(uint256(uint32((x))));
-    _keyTuple[1] = bytes32(uint256(uint32((y))));
+    _keyTuple[0] = bytes32(uint256(int256(x)));
+    _keyTuple[1] = bytes32(uint256(int256(y)));
 
     StoreSwitch.setField(_tableId, _keyTuple, 0, abi.encodePacked((value)));
   }
@@ -98,8 +98,8 @@ library TilePlayer {
   /** Set value (using the specified store) */
   function set(IStore _store, int32 x, int32 y, address value) internal {
     bytes32[] memory _keyTuple = new bytes32[](2);
-    _keyTuple[0] = bytes32(uint256(uint32((x))));
-    _keyTuple[1] = bytes32(uint256(uint32((y))));
+    _keyTuple[0] = bytes32(uint256(int256(x)));
+    _keyTuple[1] = bytes32(uint256(int256(y)));
 
     _store.setField(_tableId, _keyTuple, 0, abi.encodePacked((value)));
   }
@@ -112,15 +112,15 @@ library TilePlayer {
   /** Encode keys as a bytes32 array using this table's schema */
   function encodeKeyTuple(int32 x, int32 y) internal pure returns (bytes32[] memory _keyTuple) {
     _keyTuple = new bytes32[](2);
-    _keyTuple[0] = bytes32(uint256(uint32((x))));
-    _keyTuple[1] = bytes32(uint256(uint32((y))));
+    _keyTuple[0] = bytes32(uint256(int256(x)));
+    _keyTuple[1] = bytes32(uint256(int256(y)));
   }
 
   /* Delete all data for given keys */
   function deleteRecord(int32 x, int32 y) internal {
     bytes32[] memory _keyTuple = new bytes32[](2);
-    _keyTuple[0] = bytes32(uint256(uint32((x))));
-    _keyTuple[1] = bytes32(uint256(uint32((y))));
+    _keyTuple[0] = bytes32(uint256(int256(x)));
+    _keyTuple[1] = bytes32(uint256(int256(y)));
 
     StoreSwitch.deleteRecord(_tableId, _keyTuple);
   }
@@ -128,8 +128,8 @@ library TilePlayer {
   /* Delete all data for given keys (using the specified store) */
   function deleteRecord(IStore _store, int32 x, int32 y) internal {
     bytes32[] memory _keyTuple = new bytes32[](2);
-    _keyTuple[0] = bytes32(uint256(uint32((x))));
-    _keyTuple[1] = bytes32(uint256(uint32((y))));
+    _keyTuple[0] = bytes32(uint256(int256(x)));
+    _keyTuple[1] = bytes32(uint256(int256(y)));
 
     _store.deleteRecord(_tableId, _keyTuple);
   }
